Introduce CommitSubDag::height
This PR introduces `CommitSubDag::height` which is a simple index of the sub dag in the commit sequence.

In addition, this PR adds few doc comments for commit related data structures and traits.

This PR also refactors recovery code by introducing `CommitObserverRecoveredState` structure, replacing a tuple of arguments passed to the commit observer - this is probably majority of code changes in this PR.