[refactor] Add BlockStore into CommitObserver and Linearizer
Right now we pass BlockStore as an argument every call to the `CommitObserver` and `Linearizer`. This was an artefact of very early days when BlockStore was not cloneable. Since it can be cloned now, it is more natural to just keep it in the CommitObserver and Linearizer instance instead of passing it throught the call chain.

It will also come useful later when we need to use BlockHandler during replay